feat: complete overhaul of eocp date logic
